# Docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Application Modernization Web App REST API",
    "description": "Backend REST API for the Application Modernization",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhot:8080/api/ver{version}/",
      "description": "Local Development",
      "variables": {
        "version": {
          "default": "1",
          "description": "The API version number"
        }
      }
    }
  ],
  "paths": {
...
    "/worklist/accounts/{patientId}/chargeCodes/{chargeCodeId}": {
      "delete": {
        "tags": [
          "WorklistAccounts"
        ],
        "summary": "Deletes Charge Code for the given account.",
        "description": "<p>Data deleted from 'ChargeCodeDetails' table.</p>",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "description": "The patient Id.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The patient Id.",
              "nullable": true
            }
          },
          {
            "name": "chargeCodeId",
            "in": "path",
            "description": "The Charge Code Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The Charge Code Id.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "WorklistAccounts"
        ],
        "summary": "Updates Charge Code data for the given account.",
        "description": "<p>Data updated in 'ChargeCodeDetails' table.</p>",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "description": "The patient Id.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The patient Id.",
              "nullable": true
            }
          },
          {
            "name": "chargeCodeId",
            "in": "path",
            "description": "The Charge Code Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The Charge Code Id.",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The Rev/CPT Code model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChargeCodeModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChargeCodeModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChargeCodeModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
    "/worklist/accounts/{patientId}/payments/{payerNumber}": {
      "delete": {
        "tags": [
          "WorklistAccounts"
        ],
        "summary": "Deletes given payment of the account.",
        "description": "<p>Data deleted from 'tblPaymentDetail' table.</p>\r\n<p>Note: Some of the field mappings clarification will be provided during development phase.</p>",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "description": "The patient Id.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The patient Id.",
              "nullable": true
            }
          },
          {
            "name": "payerNumber",
            "in": "path",
            "description": "The payer number.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The payer number.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "WorklistAccounts"
        ],
        "summary": "Gets Payment Detail for the given account.",
        "description": "<p>Data retrieved from 'tblPaymentDetail' table.</p>\r\n<p>Note: Some of the field mappings clarification will be provided during development phase.</p>",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "description": "The patient Id.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The patient Id.",
              "nullable": true
            }
          },
          {
            "name": "payerNumber",
            "in": "path",
            "description": "The payer number.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The payer number.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentDetailModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
...
    "/worklist/accounts/{patientId}/revCpt": {
      "get": {
        "tags": [
          "WorklistAccounts"
        ],
        "summary": "Gets Rev/CPT Code items for the given account.",
        "description": "<p>Data retrieved from 'RevCPTDetail' table.</p>\r\n<p>Note: Some of the field mappings clarification will be provided during development phase.</p>",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "description": "The patient Id.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The patient Id.",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevCptCodeModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Gets or sets the message.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "An API error Model."
      },
...
      "CreateChargeCodeModel": {
        "required": [
          "chargeCode"
        ],
        "type": "object",
        "properties": {
          "chargeCode": {
            "type": "string"
          },
          "revenueCode": {
            "type": "string",
            "nullable": true
          },
          "deniedCharges": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "cptCode": {
            "type": "string",
            "nullable": true
          },
          "nonCoveredCharges": {
            "type": "number",
            "format": "double"
          },
          "transactionId": {
            "type": "string",
            "nullable": true
          },
          "units": {
            "type": "number",
            "format": "double"
          },
          "nonBilledCharges": {
            "type": "number",
            "format": "double"
          },
          "serviceDate": {
            "type": "string",
            "format": "date-time"
          },
          "charges": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "The create Charge Code model."
      },
...
    },
    "securitySchemes": {
      "jwt_bearer": {
        "type": "http",
        "description": "Authentication using JWT Token in header.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}